tutorials/gene-expression-analysis-pca-tsne-umap/requirements.txt
pandas==2.0.3
numpy==1.25.2
matplotlib==3.7.1
scikit-learn==1.3.0
umap-learn==0.5.2
